Remove btcd stuff
