+++ src/renderer/App/RpcIssueDialog/RpcIssueDialog.tsx
@@ -1,7 +1,3 @@
-import { FeatureFlags } from "_f/FeatureFlags";
-import { AwaitBtcd } from "./AwaitBtcd/AwaitBtcd";
 import { FixBcoreConnection } from "./FixBcoreConnection/FixBcoreConnection";
 
-export const RpcIssueDialog = FeatureFlags.useBcore
-  ? FixBcoreConnection
-  : AwaitBtcd;
+export const RpcIssueDialog = FixBcoreConnection;
